+++ lib/src/core/index.dart
@@ -2,3 +2,4 @@ export '../presentation/views/index.dart';
 export '../models/index.dart';
 export 'enums/index.dart';
 export 'utils/index.dart';
+export 'services/index.dart';