+++ lib/src/core/constants/liveness_detection_step_constant.dart
@@ -25,4 +25,8 @@ List<LivenessDetectionStepItem> stepLiveness = [
     step: LivenessDetectionStep.smile,
     title: "Smile",
   ),
+  LivenessDetectionStepItem(
+    step: LivenessDetectionStep.mouthOpen,
+    title: "Open Your Mouth",
+  ),
 ];